fix: swap FieldLogger for Loggers (cleaner log-outputs)
