+++ api/client_test.go
@@ -101,7 +101,7 @@ func genTestAPICli(t *testing.T) (*Client, context.Context, context.CancelFunc) 
 		Endpoint:     devnetBeaconAPI,
 		StateTimeout: StateTimeout,
 		QueryTimeout: QueryTimeout,
-		Logger:       log.WithFields(log.Fields{}),
+		Logger:       log.New(),
 	}
 
 	httpCli, err := NewClient(cfg)
